requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
confluent-kafka==1.9.2
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
msgpack==1.0.4
networkx==2.8.8
numpy==1.24.1
packaging==22.0
pandas==1.5.2
patsy==0.5.3
Pillow==9.3.0
progress==1.6
psycopg2-binary==2.9.5
py-radix==0.10.0
pyfsdb==2.1.2
pygeoip==0.3.2
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==3.14.2
python-socketio==4.6.1
pytz==2022.7
requests==2.28.1
ripe.atlas.cousteau==1.5.1
scipy==1.9.3
six==1.16.0
statsmodels==0.13.5
urllib3==1.26.13
websocket-client==1.4.2
